code/backend/requirements.txt
# Web framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Caching==2.1.0

# Database
SQLAlchemy==2.0.21
alembic==1.12.0

# Auth & security
PyJWT==2.8.0
Werkzeug==2.3.7
bcrypt==4.0.1
cryptography==41.0.4

# Data & numerics
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.3
statsmodels==0.14.0

# Financial data
yfinance==0.2.28
requests==2.31.0

# ML (optional – graceful fallback if absent)
joblib==1.3.2

# Server
gunicorn==21.2.0
python-dotenv==1.0.0

# Testing
pytest==7.4.2
pytest-flask==1.2.1
pytest-cov==4.1.0
